refactor: refactor dtos (#12)
