+++ pom.xml
@@ -60,6 +60,12 @@
             <scope>runtime</scope>
         </dependency>
 
+        <!-- Model mapper -->
+        <dependency>
+            <groupId>org.modelmapper</groupId>
+            <artifactId>modelmapper</artifactId>
+            <version>3.1.0</version>
+        </dependency>
 
     </dependencies>
