+++ src/main/java/org/instituteatri/deep/controller/InvestigationController.java
@@ -1,4 +1,4 @@
-package org.instituteatri.deep.controllers;
+package org.instituteatri.deep.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
