+++ src/main/java/org/instituteatri/deep/service/DocumentsAndEvidencesService.java
@@ -1,4 +1,8 @@
 package org.instituteatri.deep.service;
 
+import org.springframework.stereotype.Service;
+
+@Service
 public class DocumentsAndEvidencesService {
+
 }
